backend/requirements.txt
Django==3.2.5
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-allauth
dj-rest-auth
urllib3
requests
djangorestframework-simplejwt
channels-redis
channels
daphne==4.1.2
Pillow==9.4.0
django-otp
qrcode[pil]
pyotp
